% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_sf.R
\name{is_sf}
\alias{is_sf}
\title{is_sf}
\usage{
is_sf(x)
}
\arguments{
\item{x}{R object to be tested}
}
\value{
Logical. The function returns T/F
}
\description{
Helper function to identify whether an object is of class 'sf'
}
