% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize_me.R
\name{optimize_me}
\alias{optimize_me}
\title{optimize_me}
\usage{
optimize_me(x, fp, anno, preds)
}
\arguments{
\item{x}{The parameter to be solved for}

\item{fp}{Integer. The total number of false positives.}

\item{anno}{Integer. The total number of truth annotations.}

\item{preds}{Integer. The total number of predictions.}
}
\value{
Numeric. Returns value of function to be minimized when solving for x
}
\description{
Internal helper function defining the function to be minimized to solve for the number of true positives and false negatives given known values of false postives, total annotations, and total predictions.
}
