% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize_vals.R
\name{optimize_vals}
\alias{optimize_vals}
\title{optimize_vals}
\usage{
optimize_vals(x, fp, anno)
}
\arguments{
\item{x}{Integer. The solution to the minimization of the 'optimize_me' function.}

\item{fp}{Integer. The total number of false positives.}

\item{anno}{Integer. The total number of truth annotations.}
}
\value{
Numeric vector containing the number of false positives, false negatives, and true positives
}
\description{
Simple function
}
