% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_image_class.R
\name{plot_image_class}
\alias{plot_image_class}
\title{plot_image_class}
\usage{
plot_image_class(
  dat1,
  dat2,
  dat3,
  dat4,
  conf.thresh,
  over1,
  over2,
  image,
  class
)
}
\arguments{
\item{dat1}{List. The list of sf files that contain annotation polygons for the selected image and class}

\item{dat2}{Vector of polygon IDs for all false positive predictions.}

\item{dat3}{Data frame listing object classes ('CLASS') and arbitrary numeric code ('INDEX')}

\item{dat4}{Vector containing image names}

\item{conf.thresh}{A numeric value from 0 to 0.99 that specifies the minimum value for including model predictions in the analysis. Predictions with confidence levels below the threshold will be eliminated from the analysis.}

\item{over1}{A numeric value from 0 to 1 that specifies the minimum proportion of a truth annotation that must be covered by a model prediction. Predictions that do not meet this threshold will be assigned a status of 'false positive'}

\item{over2}{A numeric value from 0 to 1 that specifies the minimum proportion of a prediction area that must overlap with a truth annotation. Predictions that do not meet this threshold will be assigned a status of 'false positive'. Note that predictions that fail to meet the 'over1' threshold, but do meet the 'over2}

\item{image}{Integer. The numeric value for the selected image to plot.}

\item{class}{Character. The name of the object class to plot.}
}
\value{
List of length 2. The first component is a list containing a list of length 3 with the geometries for truth annotations, false positive, and true positive predictions. THe second component is a bounding box for proper plot X and Y limits.
}
\description{
Helper function to corral data for plotting in the app.
}
