% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vvipr-package.R
\docType{package}
\name{vvipr_package}
\alias{vvipr_package}
\title{Verify VIAME Predictions}
\description{
Verify VIAME Predictions
}
\details{
This package contains functions designed to assess predictions from trained classification models  that were developed in VIAME.
}
\seealso{
\url{https://github.com/us-amlr/vvipr}
}
\author{
Jefferson Hinke \email{jefferson.hinke@noaa.gov}
}
\keyword{package}
